pyproject.toml
[project]
name = "job-board-backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.3",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.7.0",
]
